Generalise platform matching.
Currently there are only "linux" and "darwin" blocks and to match a CPU
architecture the "arch" attribute must be specified. This is not
obvious.

This adds a "platform" block that accepts a set of platform attributes
to match against, such as CPU, OS, perhaps libc version in the future,
and so on.

eg.

    platform linux amd64 { ... }
    platform darwin { ... }