Late bind XDG_CACHE_HOME and HOME in ~/bin/hermit
When hermit creates ~/bin/hermit, ensure that `$XDG_CACHE_HOME` and
`$HOME` are not expanded in it, but rather use those variables. This
allows the user to change XDG_CACHE_HOME/HOME and have hermit adapt to
those changes, and also accommodates different contents of those
variables on a NFS shared home directory or similar.

It also fixes a subtle bug where the installer script downloaded would
run and put hermit in $HERMIT_EXE but then go to run it from a hardcoded
location that may be different to $HERMIT_EXE.