Add a "prompt" option to leave the prompt alone
Add an option (both a flag and a .hermit.hcl field) to tell hermit not
to make changes to the prompt when activating an environment.

The "prompt" option is an enum with the values "env", "short", and
"none". "env" is the same as the current prompt - the environment name
and hermit icon in the prompt. "short" is the same as the short-prompt
option - just the hermit icon in the prompt. "none" prevents hermit from
changing the prompt at all.

This subsumes the short-prompt option. As a CLI flag, it has been made
hidden, but is still present for backward compatibility. If both
"prompt" and "short-prompt" are set, "short-prompt" will override
whatever has been set for "prompt".