Reproducible (r10e) hermit executables
We implement a method to deterministically build hermit executables for
{linux, darwin} on {amd64, arm64}.

At a high level, we build the hermit binaries inside a Docker container,
based on a nixOS image and nixpkgs 21.11. This allows us to fix the
build environment and build toolchain.  We also use the `-trimpath`
option in `go build`, to increase the determinism of the builds.

The Docker build can be run in CI by a machine and on a laptop by a
human being.

To test it out, try

    make r10e-build VERSION=testversion CHANNEL=testchannel

We also add a CI step for reproducible hermit builds.