GHA workflow for r10e builds test
Move the reproducible builds integration test to a separate workflow
that has different triggers than the normal CI workflow. The test is to
help catch regressions in the r10e build process. Because a reproducible
build takes a significantly longer time to complete, we move it off the
hot CI path. The workflow is triggered by changes in a handful files,
and runs on a daily schedule.