fix: use the right installer URL for the canary channel (#257)
* fix: use the right installer URL for the canary channel

- We make changes for the generated bin/hermit to correctly identity the
download URL for the canary channel. This part was overlooked in an
earlier implementation.
- The `InstallScriptSHAs` map is manually populated for the "canary" and
"stable" channels, respectively, in `env.go`.
- Update `ScriptSHAs` for the updated `bin/hermit` scripts, for "canary"
and "stable" channels.
- Add `InstallerSHA256Sums` map to the `Config` struct of the main Hermit app.
- Update GHA CI step for sanity check of SHA256 sums of install script.
- Fix a buggy integration test.

* Add installer-sha-256 command

This allows us to avoid inspecting the Hermit source for the CI test to
validate checksum coherence in Hermit and in geninstaller. It makes the
CI test more robust and less inelegant.

* Make installer sums app configurable during init

The default configuration for `InstallerSHA256Sum` is
`hermit.InstallScriptSHAs`. This value now can be overridden by the main
application.

* Fix test: envfixture

Provide an empty map as the extra input argument needed by hermit.Init().

* Run gofmt -s on env.go

Reformatting so it's conformant.