Generate ScriptSHAs from an embedded file
Instead of hardcoding known script SHA-256 digest values of `bin/hermit`
and `bin/activate-hermit`, we now have them read from an embedded file
named `script.sha256`. This is to make it easier to automate the release
process. Note that `script.sha256` is supposed to be an append-only
file.

Add a subcommand `script-sha` to print out ScriptSHAs known to Hermit.

Add a CI integration testing step to make sure file `script.sha` contains
the sha256 values corresponding to the current source code revision.

Also made an opportunistic change in `it/common.sh` to make
`fakeRelease` able to work for any channel.