Refactor: remove redundant flag, move file around
We remove the `--schema` flag from the `gen-installer` command, for it's
no longer perceived to be useful.

We move the tempate file `install.sh.tmpl` to the `files/` directory for
consistency. Due to a golang constraint (cf.
https://github.com/golang/go/issues/46056), we created a package `files`
to make the installer template available to other packages.

Thank @alecthomas and @jvmakine for their feedback.