Add hermit subcommand gen-installer (#273)
Add hermit gen-installer subcommand. It prints the SHA-256 digest of the generated installer script to stdout.

We also perform a cross-check in CI to make sure `cmd/geninstaller` and
`hermit gen-installer` produce identical installer scripts.

We remove the `--schema` flag from the `gen-installer` command, for it's
no longer perceived to be useful.

We move the tempate file `install.sh.tmpl` to the `files/` directory for
consistency. Due to a golang constraint (cf.
https://github.com/golang/go/issues/46056), we created a package `files`
to make the installer template available to other packages.