Verify installer script checksum when bootstrapping (#276)
* Add hermit subcommand gen-installer

Add hermit gen-installer subcommand. It prints the SHA-256 digest of the generated installer script to stdout.

* Use hermit gen-installer to generate installer script

We also perform a cross-check in CI to make sure `cmd/geninstaller` and
`hermit gen-installer` produce identical installer scripts.

* Verify installer script checksum when bootstrapping

We verify the downloaded install script's SHA-256 value against a known
checksum in bin/hermit. This ensures the integrity of the installer
when bootstrapping hermit.

We version (via file name suffix) the install script by the script's
short hash value, so that an install script can be uniquely identified
by its version / file name. The corresponding bin/hermit is generated
with hermit init for a project.

Note that an install script's "version" is unrelated to the version of
the hermit executable.

* Revert last three commits

Part of conflict resolution.

* Cherry pick installer script checksum changes

Second step in conflict resolution. Also made some minor updates.

* Revert "System install paths as a CLI parameter"

This reverts commit 43c8037603886d3c56255da6543ede64575c70ec.