feat: add "package-index" command
This dumps the fully resolved list of all versions of all packages. This
can be used for static package indexes, but is initially intended for
use with Renovate.

One aspect I'm not 100% happy about is that this promotes what was
previously just a public part of the Go API, into a permanent part of
the Hermit "persistent" interface (similar to the bin dir, config files,
etc.). But I don't have a better idea other than duplicating the entire
structure 1:1.