Support autoversion package sha256 verification
This change makes it possible to do package sha256 check during package
installation for autoversioned packages. We do so by using an atrribute
named `sha256sums` in the package manifest file, which is a map of
format

    {
     source1: source1_sha256,
     source2: source2_sha256
    }

e.g,

    sha256sums = {
      "https://github.com/syncom/sf-signer/releases/download/v0.0.0/sfsigner-darwin-x86_64": "3e8692fa751983b75b11392968b6244e48b8b56a3363257286840ac35a696c06",
      "https://github.com/syncom/sf-signer/releases/download/v0.0.0/sfsigner-linux-x86_64": "7cfd5c00b8bc7215bc99972017da1837c514856db6d72b8e44e9136cb4fe9054"
    }

When the package sha256 checksums in the `sha256` and `sha256sums`
attribtes do not agree, the `sha256` attribute takes precedence over the
`sha256sums` attribute, for package integrity verification.