Record the hashes for the installed packages.
This change adds a new sub command to manifest_cmd. The name
is lock_digests.

It will look at the installed packages and calculate the shasum.
It will then write down manifest files in:
`bin/lockDigests`

These files will contain the sha256sums map populated with the hash
of the installed packages.

The user can review the changes and commit the files locally to enforce
the check in future.

Note:
1. It does not exhaustively download all the permutations. Only the
   current platform and the current version.
2. This probably needs to be made platform independent.
3. Autoversion updates are not yet included.

This commit also has some prep work for changes to to the
install/update flow.