fix: support text/x-shellscript + packages that can't be extracted
Previously, shell scripts were not correctly detected, so for example
ktlint could not be packages as it is a shell wrapper around a JAR file.
Shell scripts are now detected automatically.

Additionally, packages distributed as eg. JAR files or zipapp files
would not work as Hermit would attempt to extract them. This can be
overridden with `dont-extract = true`. These kinds of packages will
require an additional wrapper script be installed as per the integration
test in this PR.

Fixes #269, #322.