feat: Git manifest sources can now specify a reference to pin to
This was in the documentation but not implemented
(as discovered in #342). `#<tag>` was supported for Git packages, so
I've factored that code out and reused it.

Added an integration test for this, along with a couple of new test helper
functions that should make these kind of tests simpler in the future.