fix: prioritize runtime deps envars
When hermit runs a binary from a package, it pulls in environment variables
from various sources, including from other installed packages and the target
package's runtime dependencies. However, the environment variables from
"other installed packages" currently overrides any environment variables with
the same name from the runtime dependencies. This seems undesirable since the
random other packages are not as important to the target package as the
runtime dependencies are. In particular if a package specifies a particular
java version as a runtime dependency, but the consumer of the package has a
different java version installed in their hermit folder, the consumer-installed
JAVA_HOME overrides the runtime dependency.

This patch changes the priority order of the environment variables slightly
so that the runtime dependencies for the target package override the environment
variables for the other unrelated packages, and then the target package's
variables override the runtime dependencies. This fixes the JAVA_HOME issue
described above, and hopefully doesn't regress any other behaviours.

Fixes #368