autoversion: Support CSS selectors
autoversion currently only supports XPath for searching for versions
in an HTML page.
This works great, but CSS selectors are arguably
more approachable for newcomers, and more widely understood.

This change adds support for CSS selectors to autoversion
using the github.com/andybalholm/cascadia package (BSD2 licensed)
for matching them.

To fit that in, this extracts the search logic into a new abstraction
called htmlMatcher with an implementation for XPath and one for CSS.

Following this change, you can do something like the following:

```hcl
auto-version {
  html {
    url = "https://example.com"
    css = "div.version"
  }
}
```

It is invalid to specify both `xpath` and `css`
in the same auto-version block.