fix: sync packages if "hermit install" fails to find a package
This occasionally breaks enabling the hermit IntelliJ plugin that installs dependencies automatically at startup.

We also skipt the test if we had synchronized previously if we are forcing a sync.
This is needed as "hermit install" does a non forced sync at the beginning, marking source as synced, and this prevents the later forced sync from running when the package is not found.