config: Support opting into GitHub token auth
Adds a new per-project configuration to hermit.hcl
allowing opting GitHub owner/repo patterns to opt into
token-based authentication instead of anonymous.
For example:

    github-auth-token {
      match = ["cashapp/*"]
    }

The option accepts any number of glob patterns,
allowing for fine-grained control over which dependencies
get authenticated requests.

This will make it possible to use Hermit with private repositories
without distributing a custom build.

The plumbing for this required a small amount of refactoring:
The new hermit.LoadEnvInfo inspects a Hermit environment
and loads the configuration, but does not open the environment.
The information is later passed to the old hermit.OpenEnv function.

This separation is necessary because cache and state are inputs to
OpenEnv, but we need to inspect the environment configuration to decide
on their inputs.

Resolves #406