github.Client: Add end-to-end tests
Adds some simple end-to-end tests for the GitHub client
to avoid breaking this with other planned changes.

This introduces a new test dependency [dnaeon/go-vcr][1],
but it's a pretty small surface area,
and go-vcr itself only depends on yaml.v3.

  [1]: https://github.com/dnaeon/go-vcr

The way the test works is pretty straightforward:

- In record mode (enabled with `go test -update`)
  actual requests are made to GitHub and recorded into testdata files.
- In replay mode (the default),
  the HTTP client only replays the previously recorded interactions,
  and rejects any new requests.
  So tests will not make actual requests to GitHub after first run.

This test exposed a small bug in GitHub client initialization:
It ignored the supplied client if a token was not provided.
This change fixes that bug as well.