fix: specify a message when acquiring the flock (#419)
This should help in debugging what is causing a flock to timeout.

I rolled my own flocking implementation which is slightly less
featureful than the old package we were using, in that it doesn't
support Window. But as Hermit doesn't support Windows anyway, this
shouldn't be an issue.

Also upgraded Go to 1.23.