fix(app): unsafe stdout buffer management (#424)
app.Main spawns a goroutine to periodically flush output buffers to
stdout and stderr. However, it does so without acquiring the log
writer's lock. Add a Sync method to UI so that we can ensure that the
mutex is acquired before flushing the buffers.

This change is more theoretical than anything. I discovered it by using
`go build -race` rather than after having observed any real symptoms. It
may or may not be related to #333.