perf: check installed packages first in ResolveVirtual
Previously, when `hermit exec` was run on a package with a `requires`
field, it would parse every known package to gather a list of candidates
to fulfill the requirement. In the vast majority of cases, the user will
already have the dependent package installed, so optimize by first
evaluating only the installed packages. This change has no appreciable
effect on the negative path where we *do* end up parsing all known
packages. The result is a speedup by about a factor of 5 to 10 on my machine.