perf(exec): improve performance
...by only parsing installed packages for their environment variables if
we are in a deactivated hermit environment. When activated, the parent
shell will have already passed the new process all environment
variables, including packages that were installed after initial
activation. Parsing the HCL configs of all installed packages introduces
some overhead. This can be non-negligible when hermit-managed
packages are used a large number of times - for instance with javac or
rustc.