fix: Revert "perf(exec): improve performance" (#442)
## 📝 Description
In a nested environment, Hermit fails to set the correct environment for
the target binary. For example, given the following directory structure:
```
repo-root/
├── bin/
│   ├── java@17
│   ├── gradle
├── foo/
│   ├── bin/
│   │   ├── java@21
```
When navigating to `repo-root/foo` and running `gradle`, Hermit installs
Java 17 but sets `JAVA_HOME` to Java 21, resulting in the following
error:
```
JAVA_HOME is set to an invalid directory: /root/.cache/hermit/pkg/openjdk@21
```
This issue leads to unexpected failures in internal builds.

## Revert Reason
This commit reverts cashapp/hermit#433 because it introduces a
regression in environment resolution. Until a proper fix is found,
reverting ensures stability for existing builds.