perf(exec): improve performance (#448)
The previous attempt in 9a9737ec668f5aa01cb8a2b83d5832654d6e4997 (#433)
had an invalid method of detecting when environment variables needed to
be parsed, so it was reverted in
cc90510721cc58923414ce1efd339b3e2ffb20b8 (#442). This attempt takes the
approach of instead checking if the binary being executed is inside the
currently activated environment. If not, then we load the environment
variables of all packages in the exec'd binary's environment. It may be
possible to get away with only parsing the package's dependencies, but
I'm not confident that would correctly handle all present use cases.