refactor: move common variable expansion to envars package
While working on an implementation for #312, I found that it was
necessary to factor variable expansion logic out of the manifest package
so that it can be used elsewhere, such as when parsing hermit.hcl. This
does not necessarily make the existing code cleaner, but it will
facilitate future changes.

Also fixes an issue where variable references without {} (e.g., $FOO) were
in some cases left unexpanded.