fix: expand version with channel name if version is not set
If a channel block does not set a version either explicitly or with a
glob, ${version} will expand to an empty string, which can result in
surprising errors. There are currently packages that have channels that
do not set the version and packages that use "${version}" in fields
outside version blocks, but no packages do both. As such, it has not yet
surfaced as an issue.