feat: add a `hermit bundle <dir> [<pkg> ...]` command
This command creates a self-contained bundle of Hermit packages in the given
directory. All binaries are symlinked into `<dir>/bin` and a `<dir>/.env`
directory is created that will initialise the environment in the same way as
the Hermit environment.

Only package envars are propagated into the bundle, not environment envars.